TestCapacityScheduler: found one more app related test
